refactor: move servers folder into internal
